Change structure. Change tests
