+++ cmd/main_test.go
@@ -6,8 +6,6 @@ import (
 	"net/http/httptest"
 	"strings"
 	"testing"
-	"time"
-	"yandex-practicum-go-shortener/cmd/shortener/storage"
 
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/require"
@@ -28,18 +26,6 @@ func TestHandlers(t *testing.T) {
 		require.Equal(t, response.StatusCode, http.StatusCreated)
 		require.NotEmpty(t, body)
 		shortLink = body
-	})
-
-	t.Run("Many requests", func(t *testing.T) {
-		startTime := time.Now()
-		gin.SetMode(gin.ReleaseMode)
-
-		c := time.After(time.Second)
-
-		rCount := testStress(t, c)
-
-		t.Logf("Complete %d requests for %f second\n", rCount, time.Since(startTime).Seconds())
-		t.Logf("Len storage: %d", storage.Len())
 	})
 
 	t.Run("Getting redirect", func(t *testing.T) {
@@ -63,24 +49,6 @@ func TestHandlers(t *testing.T) {
 		defer response.Body.Close()
 		require.Equal(t, response.StatusCode, http.StatusNotFound)
 	})
-}
-
-func testStress(t *testing.T, c <-chan time.Time) int {
-	i := 1
-	for {
-		select {
-		case <-c:
-			return i
-		default:
-			reader := strings.NewReader(testURL)
-			request := httptest.NewRequest(http.MethodPost, "/", reader)
-			response, body := testRequest(t, server, request)
-			defer response.Body.Close()
-			require.Equal(t, response.StatusCode, http.StatusCreated)
-			require.NotEmpty(t, body)
-			i++
-		}
-	}
 }
 
 func testRequest(t *testing.T, server *gin.Engine, request *http.Request) (*http.Response, string) {
